refactor: rename batch test script, update programmatic interface and documentation
